+++ NEWS.md
@@ -11,6 +11,7 @@
   cannot be validated. This allows deployment when using archived CRAN
   packages, or when using packages installed from source that are available on
   the server. (#508)
+* Err when the app-mode cannot be inferred; seen with empty directories/file-sets (#512)
 
 ## 0.8.17